refactor: move determineCurrentOverlayId function into reducer file
